+++ src/main/java/principal/model/Usuario.java
@@ -13,7 +13,7 @@ public class Usuario extends Pessoa {
 
 	}
 
-	public Usuario(String nome, Integer idade, Integer telefone, String email, String senha) {
+	public Usuario(String nome, Integer idade, Long telefone, String email, String senha) {
 		super(nome, idade, telefone, email, senha);
 	}
